move gha_annotation to test_utils
