+++ tests/tt_metal/CMakeLists.txt
@@ -1,5 +1,6 @@
 add_subdirectory(distributed)
 add_subdirectory(tt_fabric)
 add_subdirectory(tt_metal)
+add_subdirectory(test_utils)
 add_subdirectory(multihost)
 add_subdirectory(tools)
